feat: added gates for the routes

+++ app/Http/Controllers/UserController.php
@@ -13,11 +13,15 @@ class UserController extends Controller
 {
     public function index()
     {
+        $this->authorize('view', 'users');
+
         return UserResource::collection(User::with('role')->paginate());
     }
 
     public function store(UserCreateRequest $request)
     {
+        $this->authorize('edit', 'users');
+
         $user = User::create(
             $request->only('firstname', 'lastname', 'email', 'role_id')
             + ['password' => \Hash::make(1234)]
@@ -28,11 +32,14 @@ class UserController extends Controller
 
     public function show($id)
     {
+        $this->authorize('view', 'users');
+
         return new UserResource(User::with('role')->find($id));
     }
 
     public function update(UserUpdateRequest $request, $id)
     {
+        $this->authorize('edit', 'users');
         $user = User::find($id);
         $user->update($request->only('firstname', 'lastname', 'email'));
 
@@ -41,6 +48,7 @@ class UserController extends Controller
 
     public function destroy($id)
     {
+        $this->authorize('edit', 'users');
         User::destroy($id);
 
         return \response(null, Response::HTTP_NO_CONTENT);
